Moved internal OrdinalAxis to ts-node.
